third_party/blink/web_tests/platform/generic/fast/forms/radio/radio-checked-pseudo-selector-expected.txt

PASS view.getComputedStyle(input1, '').getPropertyValue('color') is "rgb(0, 128, 0)"
PASS input1.checked is true
PASS successfullyParsed is true

TEST COMPLETE
